pyproject.toml
[tool.poetry]
name = "finbot-django"
version = "0.1.0"
description = ""
authors = ["Rus Esakov <eucle@ya.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = ">=4.1,<5.0"
django-crispy-forms = "^2.1"
crispy-bootstrap4 = "^2023.1"
psycopg = {extras = ["binary", "pool"], version = "^3.1.12"}
django-admin-rangefilter = "^0.11.2"
environs = "^9.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
